#22 Change DecreasePercent to decimal for fractional percentages
Updated DecreaseCapitalDto to use decimal? for both DecreasePercent and
CapitalDecreaseValue properties to support fractional percentages instead
of long integers. This allows the system to handle percentage values with
decimal precision as required by Iranian financial reporting standards.

Also fixed System.IO.Path usage in ExtraAssemblyTestData helper class to
avoid ambiguity with the Path static class from System.IO namespace.

+++ src/Application/Fundamental.Application/Codals/Dto/FinancialStatements/ManufacturingCompanies/ExtraAssembly/V1/DecreaseCapitalDto.cs
@@ -8,10 +8,10 @@ namespace Fundamental.Application.Codals.Dto.FinancialStatements.ManufacturingCo
 public class DecreaseCapitalDto
 {
     [JsonProperty("capitalDecreaseValue")]
-    public long? CapitalDecreaseValue { get; set; }
+    public decimal? CapitalDecreaseValue { get; set; }
 
     [JsonProperty("decreasePercent")]
-    public long? DecreasePercent { get; set; }
+    public decimal? DecreasePercent { get; set; }
 
     [JsonProperty("isAccept")]
     public bool IsAccept { get; set; }
